Add new implementation for in-toto 0.1.0
This is the fist set of changes for the new implementation of in-toto spec. This includes:

* Moving of existing functionality into legacy package
* Creating a set of base model classes
* Implementing validators
* Implementing helper class to start validation process and transform to JSON
* Set of tests for the above
