+++ README.md
@@ -9,38 +9,70 @@ its current limitations.
 
 ## installation
 
-This library is intended to be used with maven buildsystem, although you can
-probably easily move it to any other if you're familiar with those. To add it
-to your mvn project edit the pom.xml file to add:
+This library is intended to be used with maven build system, although you can
+probably easily move it to any other if you're familiar with those. To add it to
+your mvn project edit the pom.xml file to add:
 
 ```xml
     ...
-    <dependency>
-      <groupId>io.github.in-toto</groupId>
-      <artifactId>in-toto</artifactId>
-      <version>0.1</version>
-      <scope>compile</scope>
-    </dependency>
+<dependency>
+  <groupId>io.github.in-toto</groupId>
+  <artifactId>in-toto</artifactId>
+  <version>0.0.2</version>
+</dependency>
     ...
 ```
 
-With it you should be able to use the library inside of your project.
+With it, you should be able to use the library inside your project.
 
-## Using the library
+## Using the new library
 
-The library exposes a series of objects and convenience methods to create,
-sign, and serialize in-toto metadata. As of now, only Link metadata is
-supported (see the Limitations section to see what exactly is supported as of
-now).
+The library exposes a new set of models used for in-toto attestations 0.1.0. If
+you wish to use the deprecated legacy Link library, please skip to the next
+section.
 
-Metadata classes are located in the `io.in_toto.models.*` namespace. You can,
-for example create a link as follows:
+The new library allows you to instantiate a Statement and populate it as
+follows:
 
 ```java
-    Link link = new Link(null, null, "test", null, null);
+Subject subject=new Subject();
+    subject.setName("curl-7.72.0.tar.bz2");
+    subject.setDigest(
+    Map.of(
+    DigestSetAlgorithmType.SHA256.toString(),
+    "d4d5899a3868fbb6ae1856c3e55a32ce35913de3956d1973caccd37bd0174fa2"));
+    Predicate predicate=new Predicate(); // Let's pretend this is an SLSA predicate
+    Statement statement=new Statement();
+    statement.set_type(StatementType.STATEMENT_V_0_1);
+    statement.setSubject(List.of(subject));
+    statement.setPredicateType(PredicateType.SLSA_PROVENANCE_V_0_1);
+    statement.setPredicate(predicate);
 ```
 
-This will create a link object that you can operate with. 
+Finally, you can use the built-in `IntotoHelper` class to validate and transform
+it into its JSON representation as follows:
+
+```java
+    String jsonStatement=IntotoHelper.validateAndTransformToJson(statement);
+```
+
+If the statement passed to the method is malformed the library will throw
+an `InvalidModelException` that will contain a message with the errors.
+
+## Using the legacy Link library
+
+The library exposes a series of objects and convenience methods to create, sign,
+and serialize in-toto metadata. As of now, only Link metadata is supported (see
+the Limitations section to see what exactly is supported as of now).
+
+Metadata classes are located in the `io.github.legacy.models.*` package. You
+can, for example create a link as follows:
+
+```java
+    Link link=new Link(null,null,"test",null,null);
+```
+
+This will create a link object that you can operate with.
 
 You can populate a link and track artifacts using the Artifact class and the
 ArtifactHash subclass. You can also use the link's convenience method:
@@ -49,54 +81,31 @@ ArtifactHash subclass. You can also use the link's convenience method:
     link.addArtifact("alice");
 ```
 
-Once the artfifact is populated, it hashes the target artifact with any of the
+Once the artifact is populated, it hashes the target artifact with any of the
 supported hashes.
 
 Finally, you can sign and dump a link by calling sign and dump respectively.
 
 ```java
-import io.github.in_toto.keys.Key;
-import io.github.in_toto.keys.RSAKey;
+
 ...
-    Key thiskey = RSAKey.read("src/test/resources/somekey.pem");
-    System.out.println("Loaded key: " + thiskey.computeKeyId());
+    Key thiskey=RSAKey.read("src/test/resources/somekey.pem");
+    System.out.println("Loaded key: "+thiskey.computeKeyId());
 
     ...
-    Link link = new Link(null, null, "test", null, null, null);
+    Link link=new Link(null,null,"test",null,null,null);
     link.addMaterialt("alice");
 
     link.sign(thiskey);
     link.dump(somelink);
 ```
 
-You can see a complete example on `src/java/io/github/in_toto/lib/App.java`.
-
-## Note on reduced feature-set
-
-in-toto java is not yet a fully compliant in-toto implementation. This
-implementation is focused on providing a stable, usable core feature set for
-its main goal. The features that we will add in the near future include:
-
-- A more user-friendly API to create and interact with metadata.
-- Layout metadata support, including full supply chain verification.
-- DSA (and possibly GPG) key support.
-- A more thorough test suite that includes integration tests.
-
-We can guarantee that the dumped link metadata passes in-toto verification on
-the [python](https://github.com/in-toto/in-toto) reference implementation when
-providing the right key.
-
-As of now, the near-future goals of this library are to be used in a Jenkins
-plugin and to support Android buildsystems. However, for any other step in the
-supply chain I *highly* recommend you use the python implementation, for it has
-more features, it's better tested and will be updated to comply wih the spec
-before this one.
-
-If you'd like to help with the development of this library, patches are
-welcome!
+You can see a complete example on `src/java/io/github/legacy/lib/App.java`.
 
 ## Acknowledgements
 
 This work was mostly driven forward by the awesome guys at
 [control-plane](https://control-plane.io). If you're interested in cloud native
 security, do check out their website.
+
+If you'd like to help with the development of this library, patches are welcome!
