--- src/main/java/io/github/in_toto/models/Signable.java
@@ -1,25 +1,0 @@
-package io.github.in_toto.models;
-
-import io.github.in_toto.lib.JSONEncoder;
-
-
-/**
- * A signable class is an abstract superclass that provides a representation method
- * to prepare for signing
- *
- */
-abstract class Signable
-    implements JSONEncoder
-{
-    /**
-     * Subclasses must define the _type field appropriately for serialization
-     */
-    protected String _type;
-
-    public Signable() {
-        this._type = getType();
-    }
-
-    public abstract String getType();
-
-}
