+++ src/main/java/io/github/legacy/models/Signable.java
@@ -1,0 +1,18 @@
+package io.github.legacy.models;
+
+import io.github.legacy.lib.JSONEncoder;
+
+/**
+ * A signable class is an abstract superclass that provides a representation method to prepare for
+ * signing
+ */
+abstract class Signable implements JSONEncoder {
+  /** Subclasses must define the _type field appropriately for serialization */
+  protected String _type;
+
+  public Signable() {
+    this._type = getType();
+  }
+
+  public abstract String getType();
+}
