--- src/test/java/io/github/in_toto/lib/AppTest.java
@@ -1,1 +1,0 @@
-package io.github.in_toto.lib;
