+++ src/test/java/io/github/legacy/lib/AppTest.java
@@ -1,0 +1,1 @@
+package io.github.legacy.lib;
